+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+* Ensure CSS is applied to nested route layouts ([#1581](https://github.com/sveltejs/sapper/pull/1581))
 * Add TypeScript types for the `preload` function ([#1468](https://github.com/sveltejs/sapper/pull/1468))
 * Fix `preload` links for exported sites ([#1578](https://github.com/sveltejs/sapper/pull/1578))
 * Support extensions with multiple dots ([#1513](https://github.com/sveltejs/sapper/issues/1513))